enables transaction management

+++ src/main/java/pl/sobczakpiotr/model/UserDao.java
@@ -1,17 +1,20 @@
 package pl.sobczakpiotr.model;
 
+import org.springframework.stereotype.Repository;
+
 import java.util.List;
 
 /**
  * @author Piotr Sobczak, created on 19-03-2018
  */
+
 public interface UserDao {
 
   User findByUserName(String userName);
 
   User findByUserEmail(String email);
 
-  User findByUserID(long id);
+  User findByUserID(int id);
 
   List<User> getAllUsers();
